ci: Reuse common.spec.js.
